ITU/requirements.txt
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.3.24
Flask-Login==0.6.2
cryptography
psycopg2
APScheduler
Flask-APScheduler